Beginning of config modularization
It's become pretty clear that this config is going to be the one that
I'm working with for a long time. To prevent things from becoming
unmaintainable, I'm going to want to being to break the config down
into multiple sections that I can modify independently. The basic
modularization added in this commit should allow me to do that
incrementally.